kafka-starter/requirements.txt
kafka-python==2.0.2
pymongo==4.3.2